requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
contourpy==1.1.0
cryptography==41.0.5
cycler==0.11.0
distlib==0.3.7
distro==1.8.0
docxcompose==1.4.0
environs==9.5.0
etelemetry==0.3.1
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.12.4
firebase-admin==6.2.0
fonttools==4.42.1
frontend==0.0.3
frozenlist==1.4.0
future==0.18.3
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
idna==3.4
importlib-resources==6.0.1
install==1.3.5
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
minio==7.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.24.4
openai==1.3.5
orjson==3.9.10
packaging==23.1
pandas==2.0.3
pathlib==1.0.1
Pillow==10.0.0
platformdirs==3.11.0
plotly==5.16.1
proto-plus==1.22.3
protobuf==4.25.1
prov==2.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
pydot==1.4.2
PyJWT==2.8.0
pymilvus==2.3.3
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
regex==2023.8.8
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tqdm==4.66.1
traits==6.3.2
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.24.6
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
zipp==3.16.2
